flowedit: Move test code from main.rs to ui_test.rs
Moved all test code (~760 lines) from flowedit/src/main.rs into a new
dedicated test module file flowedit/src/ui_test.rs. This reduces the
size of main.rs and improves test organization. The nested mod ui
structure was flattened so all helpers and tests are at the same level.

All 141 tests continue to pass.

Part of #2580 (flowedit automated UI testing).
